Caching is now disabled by default
Before this commit, the data was usually cached in /tmp/nodes.json.
The fresh data was only used if '-u' was specified.

This behaviour has now changed and the flag '-u' is no longer
available. Instead, we use fresh data by default and introduce
a flag '-c', so the user can explicitly enable caching if desired.